// frontend/src/api/generated/activity/activity.ts
export type listActivityApiV1ActivityGetResponse200 = {
  data: LimitOffsetPageTypeVarCustomizedActivityEventRead;
  status: 200;
};

export type listActivityApiV1ActivityGetResponse422 = {
  data: HTTPValidationError;
  status: 422;
};

export type listActivityApiV1ActivityGetResponseSuccess =
  listActivityApiV1ActivityGetResponse200 & {
    headers: Headers;
  };
export type listActivityApiV1ActivityGetResponseError =
  listActivityApiV1ActivityGetResponse422 & {
    headers: Headers;
  };

export type listActivityApiV1ActivityGetResponse =
  | listActivityApiV1ActivityGetResponseSuccess
  | listActivityApiV1ActivityGetResponseError;